picoxcell: support the AXI bus error snoopers
The picoXcell devices have a number of peripherals that may perform
master accesses to the AXI bus, but if incorrectly programmed can
generate a bus error. The AXI2CFG block has bus snoopers that monitor
the bus for errors from other masters and raises an interrupt if
detected. These errors are typically irrecoverable so panic if we see
them.

Some devices have separate read and write error interrupts and each
device variant has a different mapping of bits so represent this as a
platform device.

Signed-off-by: Jamie Iles <jamie@jamieiles.com>