picoxcell: add support for uart based UICC
The UART based UICC controller allows userspace to implement a UICC
stack.  This adds a new UART which may be used to send and receive data
to/from the UICC device and adds a clk_en attribute to the device that
can be used to turn the clock on and off.  Note that the configuration
means that data sent to the device will be echoed back to the serial
port and needs to be removed in userspace.

Signed-off-by: Jamie Iles <jamie@jamieiles.com>