picoxcell: add support for the clk framework
Implement the clk operations required in linux/clk.h and provide the
basic infrastructure for populating the clock tree from the device tree.
Clock data is obtained from the device tree and the clock tree is built
up after reading all of the clocks.

This implements a picoxcell specific struct clk, but hopefully in a way
which can be ported to a common struct clk later.

Signed-off-by: Jamie Iles <jamie@jamieiles.com>