picoxcell: add a binding for the gated clock type
The gated clock runs at a fixed rate and can be disabled by writing to
the clk gating register in the AXI2CFG.  Add a device tree binding so
that we can populate these clocks from the device tree.  Gated clocks
are currently present in pc3x3 and pc30xx devices.

Signed-off-by: Jamie Iles <jamie@jamieiles.com>