of: add clock providers
Based on work by Ben Herrenschmidt, this patch adds an of_clk_get
function to allow platforms to retrieve clock data from the device tree.

Platform register a provider through of_clk_add_provider, which will be
called when a device references the provider's OF node for a clock
reference.

Signed-off-by: Jeremy Kerr <jeremy.kerr@canonical.com>
[grant.likely@secretlab.ca: fix Kconfig conflict]
Signed-off-by: Grant Likely <grant.likely@secretlab.ca>