picoxcell: add reset support
If we have a watchdog timer present then use that as it should reset the
whole SoC.  If not then we can fall back to a cpu soft reset by jumping
back to the reset vector.

In future we should be able to support platforms that use a GPIO for
board level reset from the device tree.

Signed-off-by: Jamie Iles <jamie@jamieiles.com>